balls

+++ rawreader_windows.go
@@ -58,7 +58,7 @@ next:
 			return 0, err
 		}
 	*/
-	_, _, err = kernel.ReadConsoleInputW.Call(stdin, uintptr(unsafe.Pointer(&irArr[0])), uintptr(1), uintptr(unsafe.Pointer(pread)))
+	_, _, err = kernel.ReadConsoleInputW.Call(stdin, uintptr(unsafe.Pointer(&irArr[0])), uint32(1), uintptr(unsafe.Pointer(pread)))
 	ir := irArr[0]
 	if err != nil {
 		fmt.Printf("\nReadConsoleInputW: err=%v ir=%v read=%v pread=%v\n", err, ir, read, pread)
